move the thunderfx related functions into dynamo
