+++ thunder/dynamo/__init__.py
@@ -1,6 +1,7 @@
-from thunder.dynamo.compiler import ThunderCompiler
+from thunder.dynamo.compiler import ThunderCompiler, ThunderCompilerGraphBenchmarking
 
 
 __all__ = [
     "ThunderCompiler",
+    "ThunderCompilerGraphBenchmarking",
 ]
